fixed duplicate of ScenarioDataParser
